jet: move `exec_jet` function into `BitMachine`
This function is 90% bit machine frame manipulation. It is better for it
to live in `bit_machine`. This lets us eliminate the last of the
`pub(crate)`s from the bit machine, and completely hide the `frame`
module.
