+++ src/jet/mod.rs
@@ -36,11 +36,9 @@ pub use init::bitcoin::Bitcoin;
 pub use init::core::Core;
 #[cfg(feature = "elements")]
 pub use init::elements::Elements;
-use simplicity_sys::c_jets::frame_ffi::{c_readBit, c_writeBit, CFrameItem};
-use simplicity_sys::c_jets::round_u_word;
+use simplicity_sys::c_jets::frame_ffi::CFrameItem;
 
 use crate::decode;
-use crate::exec::BitMachine;
 use crate::jet::type_name::TypeName;
 use crate::merkle::cmr::Cmr;
 use crate::{BitIter, BitWriter};
@@ -95,73 +93,6 @@ pub trait Jet: Copy + Eq + Ord + Hash + std::fmt::Debug + std::fmt::Display {
 
     /// Obtain the FFI C pointer for the jet.
     fn c_jet_ptr(&self) -> &dyn Fn(&mut CFrameItem, CFrameItem, &Self::CJetEnvironment) -> bool;
-
-    /// Execute the jet on the Bit Machine, using the given environment.
-    fn exec<'env>(
-        &self,
-        mac: &mut BitMachine,
-        env: &'env Self::Environment,
-    ) -> Result<(), JetFailed>
-    where
-        Self::CJetEnvironment: 'env,
-    {
-        // Sanity Check: This should never really fail, but still good to do
-        if !simplicity_sys::c_jets::sanity_checks() {
-            return Err(JetFailed);
-        }
-        let src_ty_bit_width = self.source_ty().to_bit_width();
-        let target_ty_bit_width = self.target_ty().to_bit_width();
-
-        let a_frame_size = round_u_word(src_ty_bit_width);
-        let b_frame_size = round_u_word(target_ty_bit_width);
-        // a_frame_size + b_frame_size must be non-zero unless it is a unit to unit jet
-        if a_frame_size == 0 && b_frame_size == 0 {
-            return Ok(());
-        }
-        let mut src_buf = vec![0usize; a_frame_size + b_frame_size];
-        let src_ptr_end = unsafe { src_buf.as_mut_ptr().add(a_frame_size) }; // A frame write
-        let src_ptr = src_buf.as_mut_ptr(); // A read frame at ptr start
-        let dst_ptr_begin = unsafe { src_buf.as_mut_ptr().add(a_frame_size) }; // B read frame at ptr begin
-        let dst_ptr_end = unsafe { src_buf.as_mut_ptr().add(a_frame_size + b_frame_size) }; // B write frame at ptr end
-
-        // For jet from type A -> B
-        // Jets execution: There is single buffer with a_frame_size + b_frame_size UWORDs
-        // ------[ A read frame     ][    B write frame  ]---
-        //       ^ src_ptr         ^src_ptr_end(dst_ptr_begin)      ^ dst_ptr_end
-        // 1. Write into C bitmachine using A write frame(= src_ptr_end)
-        // Precondition satisfied: src_ptr_end is one past the end of slice of UWORDs for A.
-        let mut a_frame = unsafe { CFrameItem::new_write(src_ty_bit_width, src_ptr_end) };
-        for _ in 0..src_ty_bit_width {
-            let bit = mac.read_bit();
-            unsafe {
-                c_writeBit(&mut a_frame, bit);
-            }
-        }
-        mac.back(src_ty_bit_width);
-
-        // 2. Execute the jet. src = A read frame, dst = B write frame
-        // Precondition satisfied: src_ptr is the start of slice of UWORDs of A.
-        let src_frame = unsafe { CFrameItem::new_read(src_ty_bit_width, src_ptr) };
-        // Precondition satisfied: dst_ptr_end is one past the end of slice of UWORDs of B.
-        let mut dst_frame = unsafe { CFrameItem::new_write(target_ty_bit_width, dst_ptr_end) };
-        let jet_fn = self.c_jet_ptr();
-        let c_env = self.c_jet_env(env);
-        let res = jet_fn(&mut dst_frame, src_frame, c_env);
-
-        if !res {
-            return Err(JetFailed);
-        }
-
-        // 3. Read the result from B read frame
-        // Precondition satisfied: dst_ptr_begin is the start of slice of UWORDs of B.
-        let mut b_frame = unsafe { CFrameItem::new_read(target_ty_bit_width, dst_ptr_begin) };
-        // Read the value from b_frame
-        for _ in 0..target_ty_bit_width {
-            let bit = unsafe { c_readBit(&mut b_frame) };
-            mac.write_bit(bit);
-        }
-        Ok(())
-    }
 }
 
 #[cfg(test)]
